+++ CHANGES.md
@@ -2,6 +2,12 @@
 
 ## bunyan 0.3.0 (not yet released)
 
+- `log.child(options[, simple])` Added `simple` boolean arg. Set `true` to
+  assert that options only add fields (no config changes). Results in a 10x
+  speed increase in child creation. See "tools/timechild.js". On my Mac, 
+  "fast child" creation takes about 0.001ms. IOW, if your app is dishing
+  10,000 req/s, then creating a log child for each request will take
+  about 1% of the request time.
 - `log.clone` -> `log.child` to better reflect the relationship: streams and
   serializers are inherited. Streams can't be removed as part of the child
   creation. The child doesn't own the parent's streams (so can't close them).